requirements.txt
aiosqlite==0.17.0
autopep8==1.6.0
iso8601==1.0.2
mypy==0.971
mypy-extensions==0.4.3
pycodestyle==2.8.0
pypika-tortoise==0.1.6
pytz==2022.1
toml==0.10.2
tomli==2.0.1
tortoise-orm==0.19.2
typing-extensions==4.3.0
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.0.1
mypy-extensions==0.4.3
paho-mqtt==1.6.1
passlib==1.7.4
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
starlette-csrf==1.4.2
tomli==2.0.1
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8
websockets==10.2
